requirements.txt
torch
transformers
datasets
faiss-cpu
sentencepiece
fastapi
uvicorn[standard]
pydantic
scikit-learn
numpy
tqdm
pandas
python-dotenv

